@@ -1,11 +1,11 @@
 source 'https://rubygems.org'
 
-gem 'rails', '~> 4.0.12'
+gem 'rails', '~> 4.2'
 gem 'rails-i18n'
 
-gem 'mysql2'
+gem 'mysql2', '~> 0.3.20'
 gem 'daemons-rails'
-gem 'redis-rails'
+gem 'redis-rails', '5.0.1'
 
 gem 'rotp'
 gem 'json'
@@ -13,9 +13,9 @@ gem 'jbuilder'
 gem 'bcrypt-ruby', '~> 3.1.2'
 
 gem 'doorkeeper', '~> 1.4.1'
-gem 'omniauth', '~> 1.2.1'
-gem 'omniauth-identity', '~> 1.1.1'
-gem 'omniauth-weibo-oauth2', '~> 0.4.0'
+gem 'omniauth'
+gem 'omniauth-identity'
+gem 'omniauth-weibo-oauth2'
 
 gem 'figaro'
 gem 'hashie'
@@ -26,7 +26,7 @@ gem 'bunny', '~> 1.2.1'
 gem 'cancancan'
 gem 'enumerize'
 gem 'datagrid'
-gem 'acts-as-taggable-on'
+gem 'acts-as-taggable-on', '~> 4.0'
 gem 'kaminari'
 gem 'paranoid2'
 gem 'active_hash'
@@ -65,16 +65,18 @@ gem 'rack-attack', '~> 3.0.0'
 gem 'easy_table'
 gem 'phonelib', '~> 0.3.5'
 gem 'twilio-ruby', '~> 3.11'
-gem 'unread', github: 'peatio/unread'
+# gem 'unread', github: 'peatio/unread'
+gem 'unread'
 gem 'carrierwave', '~> 0.10.0'
 gem 'simple_captcha2', require: 'simple_captcha'
 gem 'rest-client', '~> 1.6.8'
 
 group :development, :test do
-  gem 'factory_girl_rails'
+  gem 'factory_bot_rails'
   gem 'faker', '~> 1.4.3'
   gem 'mina'
-  gem 'mina-slack', github: 'peatio/mina-slack'
+  # gem 'mina-slack', github: 'peatio/mina-slack'
+  gem 'mina-slack', github: 'Apollo-IT/mina-slack'
   gem 'meta_request'
   gem 'better_errors'
   gem 'binding_of_caller'
@@ -84,6 +86,7 @@ group :development, :test do
   gem 'timecop'
   gem 'dotenv-rails'
   gem 'rspec-rails'
+  gem 'rspec-its'
   gem 'byebug'
 end
 
