@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :account do
     locked { "0.0".to_d }
     balance { "100.0".to_d }
@@ -9,4 +9,3 @@ FactoryGirl.define do
     end
   end
 end
-
