@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :two_factor do
     member
 
@@ -10,4 +10,3 @@ FactoryGirl.define do
   factory :two_factor_app, class: TwoFactor::App, parent: :two_factor, traits: [:activated]
   factory :two_factor_sms, class: TwoFactor::Sms, parent: :two_factor, traits: [:activated]
 end
-
