@@ -54,7 +54,7 @@ RSpec.configure do |config|
   #     --seed 1234
   config.order = "random"
 
-  config.include FactoryGirl::Syntax::Methods
+  config.include FactoryBot::Syntax::Methods
 
   config.before(:suite) do
     DatabaseCleaner.strategy = :deletion
@@ -74,4 +74,3 @@ RSpec.configure do |config|
     DatabaseCleaner.clean
   end
 end
-
